+++ tests/Pest.php
@@ -2,8 +2,6 @@
 
 declare(strict_types=1);
 
-use Aristonis\BlogManager\Tests\ApiEnabledTestCase;
 use Aristonis\BlogManager\Tests\TestCase;
 
 uses(TestCase::class)->in('Feature', 'Unit');
-uses(ApiEnabledTestCase::class)->in('Api');